+++ CHANGELOG.md
@@ -6,11 +6,14 @@ All notable changes to this project will be documented in this file. The format 
 ### Added
 - Add optional output location for `FieldContainer.extract(out=None)`.
 - Add `Mesh.update(callback=None)`. This is especially useful if the points array of a mesh is changed and an already existing instance of a region has to be reloaded: `Mesh.update(points=new_points, callback=region.reload)`.
+- Add `ax = FieldContainer.imshow()` which acts as a wrapper on top of the `img = FieldContainer.screenshot(filename=None)` method. The image data is passed to a matplotlib figure and the `ax` object is returned.
+- Add `ax = Mesh.imshow()` which acts as a wrapper on top of the `img = Mesh.screenshot(filename=None)` method. The image data is passed to a matplotlib figure and the `ax` object is returned.
 
 ### Changed
 - Pass optional keyword-arguments in `math.dot(**kwargs)` to the underlying einsum-calls.
 - Enhance the performance of `math.cdya()` by reducing the number of (intermediate) arrays to be created from 4 to 2.
 - Use fixed output locations for the extracted field-gradients and the integrated stiffness matrices in `SolidBody` and `SolidBodyNearlyIncompressible`. This enhances the performance.
+- Change default filename in `Mesh.screenshot()` from `filename="field.png"` to `filename="mesh.png"`.
 
 ### Removed
 - Do not invoke `pyvista.start_xvfb()` on a posix-os. If required, run it manually.